Reverted back to previous itteration with no api functions file, only sql
